fixed file path errors
-Paths that worked when running in jupyter notebook no longer were. (changed to ./ relative paths)
-Discriminator network dimensions were off. (fixed but left previous commented out)
-matrix was checked in matrix_to_wav but broke with tensors. ( Added check for != None instead)
- Mel spectrograms were using different functions. ( Changed so that both use same method with previous commented out incase we decide to use the other)

Current issue: enters into a loop with only the one file and 1 batch? I had to increase the number of customers for a test so simulations are also taking too long.

+++ experimenting with adjsims/simulation_v3.py
@@ -20,6 +20,8 @@ import math
 import time
 import logging
 
+import os
+
 class FlowBranchOperator:
     def __init__(self, probabilities, children=None, origin=None):
         """
@@ -315,8 +317,9 @@ class Sim:
                 filename = log_path + "simulation.log"
             else:
                 filename = log_path + log_name
-            # delete old log file
-            open(filename, 'w').close()
+            if os.path.exists(filename):
+                # delete old log file
+                open(filename, 'w').close()
             # create logger
             logging.basicConfig(filename=filename, filemode='w', level=logging.INFO)
